+++ models/experimental/detr3d/reference/model_3detr.py
@@ -1,0 +1,349 @@
+# SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+
+# SPDX-License-Identifier: Apache-2.0
+
+import torch
+import torch.nn as nn
+from models.experimental.detr3d.reference.model_utils import (
+    FurthestPointSampling,
+    BoxProcessor,
+)
+import math
+import numpy as np
+from functools import partial
+
+import torch
+from models.experimental.detr3d.reference.transformer import (
+    TransformerDecoderLayer,
+    TransformerDecoder,
+    TransformerEncoderLayer,
+    TransformerEncoder,
+    MaskedTransformerEncoder,
+)
+from models.experimental.detr3d.reference.position_embedding import PositionEmbeddingCoordsSine
+from models.experimental.detr3d.reference.helpers import GenericMLP
+from models.experimental.detr3d.reference.pointnet2_modules import PointnetSAModuleVotes
+
+
+class Model3DETR(nn.Module):
+    """
+    Main 3DETR model. Consists of the following learnable sub-models
+    - pre_encoder: takes raw point cloud, subsamples it and projects into "D" dimensions
+                Input is a Nx3 matrix of N point coordinates
+                Output is a N'xD matrix of N' point features
+    - encoder: series of self-attention blocks to extract point features
+                Input is a N'xD matrix of N' point features
+                Output is a N''xD matrix of N'' point features.
+                N'' = N' for regular encoder; N'' = N'//2 for masked encoder
+    - query computation: samples a set of B coordinates from the N'' points
+                and outputs a BxD matrix of query features.
+    - decoder: series of self-attention and cross-attention blocks to produce BxD box features
+                Takes N''xD features from the encoder and BxD query features.
+    - mlp_heads: Predicts bounding box parameters and classes from the BxD box features
+    """
+
+    def __init__(
+        self,
+        pre_encoder,
+        encoder,
+        decoder,
+        dataset_config,
+        encoder_dim=256,
+        decoder_dim=256,
+        position_embedding="fourier",
+        mlp_dropout=0.3,
+        num_queries=256,
+    ):
+        super().__init__()
+        self.pre_encoder = pre_encoder
+        self.encoder = encoder
+        if hasattr(self.encoder, "masking_radius"):
+            hidden_dims = [encoder_dim]
+        else:
+            hidden_dims = [encoder_dim, encoder_dim]
+        self.encoder_to_decoder_projection = GenericMLP(
+            input_dim=encoder_dim,
+            hidden_dims=hidden_dims,
+            output_dim=decoder_dim,
+            norm_fn_name="bn1d",
+            activation="relu",
+            use_conv=True,
+            output_use_activation=True,
+            output_use_norm=True,
+            output_use_bias=False,
+        )
+        self.pos_embedding = PositionEmbeddingCoordsSine(d_pos=decoder_dim, pos_type=position_embedding, normalize=True)
+        self.query_projection = GenericMLP(
+            input_dim=decoder_dim,
+            hidden_dims=[decoder_dim],
+            output_dim=decoder_dim,
+            use_conv=True,
+            output_use_activation=True,
+            hidden_use_bias=True,
+        )
+        self.decoder = decoder
+        self.build_mlp_heads(dataset_config, decoder_dim, mlp_dropout)
+
+        self.num_queries = num_queries
+        self.box_processor = BoxProcessor(dataset_config)
+        self.furthest_point_sample = FurthestPointSampling()
+
+    def build_mlp_heads(self, dataset_config, decoder_dim, mlp_dropout):
+        mlp_func = partial(
+            GenericMLP,
+            norm_fn_name="bn1d",
+            activation="relu",
+            use_conv=True,
+            hidden_dims=[decoder_dim, decoder_dim],
+            dropout=mlp_dropout,
+            input_dim=decoder_dim,
+        )
+
+        # Semantic class of the box
+        # add 1 for background/not-an-object class
+        semcls_head = mlp_func(output_dim=dataset_config.num_semcls + 1)
+
+        # geometry of the box
+        center_head = mlp_func(output_dim=3)
+        size_head = mlp_func(output_dim=3)
+        angle_cls_head = mlp_func(output_dim=dataset_config.num_angle_bin)
+        angle_reg_head = mlp_func(output_dim=dataset_config.num_angle_bin)
+
+        mlp_heads = [
+            ("sem_cls_head", semcls_head),
+            ("center_head", center_head),
+            ("size_head", size_head),
+            ("angle_cls_head", angle_cls_head),
+            ("angle_residual_head", angle_reg_head),
+        ]
+        self.mlp_heads = nn.ModuleDict(mlp_heads)
+
+    def get_query_embeddings(self, encoder_xyz, point_cloud_dims):
+        query_inds = self.furthest_point_sample(encoder_xyz, self.num_queries)
+        query_inds = query_inds.long()
+        query_xyz = [torch.gather(encoder_xyz[..., x], 1, query_inds) for x in range(3)]
+        query_xyz = torch.stack(query_xyz)
+        query_xyz = query_xyz.permute(1, 2, 0)
+
+        # Gater op above can be replaced by the three lines below from the pointnet2 codebase
+        # xyz_flipped = encoder_xyz.transpose(1, 2).contiguous()
+        # query_xyz = gather_operation(xyz_flipped, query_inds.int())
+        # query_xyz = query_xyz.transpose(1, 2)
+        pos_embed = self.pos_embedding(query_xyz, input_range=point_cloud_dims)
+        query_embed = self.query_projection(pos_embed)
+        return query_xyz, query_embed
+
+    def _break_up_pc(self, pc):
+        # pc may contain color/normals.
+
+        xyz = pc[..., 0:3].contiguous()
+        features = pc[..., 3:].transpose(1, 2).contiguous() if pc.size(-1) > 3 else None
+        return xyz, features
+
+    def run_encoder(self, point_clouds):
+        xyz, features = self._break_up_pc(point_clouds)
+        pre_enc_xyz, pre_enc_features, pre_enc_inds = self.pre_encoder(xyz, features)
+        # xyz: batch x npoints x 3
+        # features: batch x channel x npoints
+        # inds: batch x npoints
+
+        # nn.MultiHeadAttention in encoder expects npoints x batch x channel features
+        pre_enc_features = pre_enc_features.permute(2, 0, 1)
+
+        # xyz points are in batch x npointx channel order
+        enc_xyz, enc_features, enc_inds = self.encoder(pre_enc_features, xyz=pre_enc_xyz)
+        if enc_inds is None:
+            # encoder does not perform any downsampling
+            enc_inds = pre_enc_inds
+        else:
+            # use gather here to ensure that it works for both FPS and random sampling
+            enc_inds = torch.gather(pre_enc_inds, 1, enc_inds.type(torch.int64))
+        return enc_xyz, enc_features, enc_inds
+
+    def get_box_predictions(self, query_xyz, point_cloud_dims, box_features):
+        """
+        Parameters:
+            query_xyz: batch x nqueries x 3 tensor of query XYZ coords
+            point_cloud_dims: List of [min, max] dims of point cloud
+                              min: batch x 3 tensor of min XYZ coords
+                              max: batch x 3 tensor of max XYZ coords
+            box_features: num_layers x num_queries x batch x channel
+        """
+        # box_features change to (num_layers x batch) x channel x num_queries
+        box_features = box_features.permute(0, 2, 3, 1)
+        num_layers, batch, channel, num_queries = (
+            box_features.shape[0],
+            box_features.shape[1],
+            box_features.shape[2],
+            box_features.shape[3],
+        )
+        box_features = box_features.reshape(num_layers * batch, channel, num_queries)
+
+        # mlp head outputs are (num_layers x batch) x noutput x nqueries, so transpose last two dims
+        cls_logits = self.mlp_heads["sem_cls_head"](box_features).transpose(1, 2)
+        center_offset = self.mlp_heads["center_head"](box_features).sigmoid().transpose(1, 2) - 0.5
+        size_normalized = self.mlp_heads["size_head"](box_features).sigmoid().transpose(1, 2)
+        angle_logits = self.mlp_heads["angle_cls_head"](box_features).transpose(1, 2)
+        angle_residual_normalized = self.mlp_heads["angle_residual_head"](box_features).transpose(1, 2)
+
+        # reshape outputs to num_layers x batch x nqueries x noutput
+        cls_logits = cls_logits.reshape(num_layers, batch, num_queries, -1)
+        center_offset = center_offset.reshape(num_layers, batch, num_queries, -1)
+        size_normalized = size_normalized.reshape(num_layers, batch, num_queries, -1)
+        angle_logits = angle_logits.reshape(num_layers, batch, num_queries, -1)
+        angle_residual_normalized = angle_residual_normalized.reshape(num_layers, batch, num_queries, -1)
+        angle_residual = angle_residual_normalized * (np.pi / angle_residual_normalized.shape[-1])
+
+        outputs = []
+        for l in range(num_layers):
+            # box processor converts outputs so we can get a 3D bounding box
+            (
+                center_normalized,
+                center_unnormalized,
+            ) = self.box_processor.compute_predicted_center(center_offset[l], query_xyz, point_cloud_dims)
+            angle_continuous = self.box_processor.compute_predicted_angle(angle_logits[l], angle_residual[l])
+            size_unnormalized = self.box_processor.compute_predicted_size(size_normalized[l], point_cloud_dims)
+            box_corners = self.box_processor.box_parametrization_to_corners(
+                center_unnormalized, size_unnormalized, angle_continuous
+            )
+
+            # below are not used in computing loss (only for matching/mAP eval)
+            # we compute them with no_grad() so that distributed training does not complain about unused variables
+            with torch.no_grad():
+                (
+                    semcls_prob,
+                    objectness_prob,
+                ) = self.box_processor.compute_objectness_and_cls_prob(cls_logits[l])
+
+            box_prediction = {
+                "sem_cls_logits": cls_logits[l],
+                "center_normalized": center_normalized.contiguous(),
+                "center_unnormalized": center_unnormalized,
+                "size_normalized": size_normalized[l],
+                "size_unnormalized": size_unnormalized,
+                "angle_logits": angle_logits[l],
+                "angle_residual": angle_residual[l],
+                "angle_residual_normalized": angle_residual_normalized[l],
+                "angle_continuous": angle_continuous,
+                "objectness_prob": objectness_prob,
+                "sem_cls_prob": semcls_prob,
+                "box_corners": box_corners,
+            }
+            outputs.append(box_prediction)
+
+        # intermediate decoder layer outputs are only used during training
+        aux_outputs = outputs[:-1]
+        outputs = outputs[-1]
+
+        return {
+            "outputs": outputs,  # output from last layer of decoder
+            "aux_outputs": aux_outputs,  # output from intermediate layers of decoder
+        }
+
+    def forward(self, inputs, encoder_only=False):
+        point_clouds = inputs["point_clouds"]
+
+        enc_xyz, enc_features, enc_inds = self.run_encoder(point_clouds)
+        enc_features = self.encoder_to_decoder_projection(enc_features.permute(1, 2, 0)).permute(2, 0, 1)
+        # encoder features: npoints x batch x channel
+        # encoder xyz: npoints x batch x 3
+
+        if encoder_only:
+            # return: batch x npoints x channels
+            return enc_xyz, enc_features.transpose(0, 1)
+
+        point_cloud_dims = [
+            inputs["point_cloud_dims_min"],
+            inputs["point_cloud_dims_max"],
+        ]
+        query_xyz, query_embed = self.get_query_embeddings(enc_xyz, point_cloud_dims)
+        # query_embed: batch x channel x npoint
+        enc_pos = self.pos_embedding(enc_xyz, input_range=point_cloud_dims)
+
+        # decoder expects: npoints x batch x channel
+        enc_pos = enc_pos.permute(2, 0, 1)
+        query_embed = query_embed.permute(2, 0, 1)
+        tgt = torch.zeros_like(query_embed)
+        box_features = self.decoder(tgt, enc_features, query_pos=query_embed, pos=enc_pos)[0]
+
+        box_predictions = self.get_box_predictions(query_xyz, point_cloud_dims, box_features)
+        return box_predictions
+
+
+def build_preencoder(args):
+    mlp_dims = [3 * int(args.use_color), 64, 128, args.enc_dim]
+    preencoder = PointnetSAModuleVotes(
+        radius=0.2,
+        nsample=64,
+        npoint=args.preenc_npoints,
+        mlp=mlp_dims,
+        normalize_xyz=True,
+    )
+    return preencoder
+
+
+def build_encoder(args):
+    if args.enc_type == "vanilla":
+        encoder_layer = TransformerEncoderLayer(
+            d_model=args.enc_dim,
+            nhead=args.enc_nhead,
+            dim_feedforward=args.enc_ffn_dim,
+            dropout=args.enc_dropout,
+            activation=args.enc_activation,
+        )
+        encoder = TransformerEncoder(encoder_layer=encoder_layer, num_layers=args.enc_nlayers)
+    elif args.enc_type in ["masked"]:
+        encoder_layer = TransformerEncoderLayer(
+            d_model=args.enc_dim,
+            nhead=args.enc_nhead,
+            dim_feedforward=args.enc_ffn_dim,
+            dropout=args.enc_dropout,
+            activation=args.enc_activation,
+        )
+        interim_downsampling = PointnetSAModuleVotes(
+            radius=0.4,
+            nsample=32,
+            npoint=args.preenc_npoints // 2,
+            mlp=[args.enc_dim, 256, 256, args.enc_dim],
+            normalize_xyz=True,
+        )
+
+        masking_radius = [math.pow(x, 2) for x in [0.4, 0.8, 1.2]]
+        encoder = MaskedTransformerEncoder(
+            encoder_layer=encoder_layer,
+            num_layers=3,
+            interim_downsampling=interim_downsampling,
+            masking_radius=masking_radius,
+        )
+    else:
+        raise ValueError(f"Unknown encoder type {args.enc_type}")
+    return encoder
+
+
+def build_decoder(args):
+    decoder_layer = TransformerDecoderLayer(
+        d_model=args.dec_dim,
+        nhead=args.dec_nhead,
+        dim_feedforward=args.dec_ffn_dim,
+        dropout=args.dec_dropout,
+    )
+    decoder = TransformerDecoder(decoder_layer, num_layers=args.dec_nlayers, return_intermediate=True)
+    return decoder
+
+
+def build_3detr(args, dataset_config):
+    pre_encoder = build_preencoder(args)
+    encoder = build_encoder(args)
+    decoder = build_decoder(args)
+    model = Model3DETR(
+        pre_encoder,
+        encoder,
+        decoder,
+        dataset_config,
+        encoder_dim=args.enc_dim,
+        decoder_dim=args.dec_dim,
+        mlp_dropout=args.mlp_dropout,
+        num_queries=args.nqueries,
+    )
+    output_processor = BoxProcessor(dataset_config)
+    return model, output_processor
